split views into shared/ui
